refactor: Enforce `LazyExpr.from_elementwise`
See https://github.com/narwhals-dev/narwhals/pull/2716#discussion_r2160078141
