+++ narwhals/_dask/expr.py
@@ -683,3 +683,4 @@ class DaskExpr(
     rank = not_implemented()  # pyright: ignore[reportAssignmentType]
     _alias_native = not_implemented()
     window_function = not_implemented()  # pyright: ignore[reportAssignmentType]
+    from_elementwise = not_implemented()
