+++ narwhals/_spark_like/namespace.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import operator
 from functools import reduce
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
 
 from narwhals._compliant import LazyNamespace, LazyThen, LazyWhen
 from narwhals._expression_parsing import (
@@ -73,31 +73,6 @@ class SparkLikeNamespace(
         else:
             return import_native_dtypes(self._implementation)
 
-    def _expr_from_callable(
-        self, func: Callable[[Iterable[Column]], Column], *exprs: SparkLikeExpr
-    ) -> SparkLikeExpr:
-        def call(df: SparkLikeLazyFrame) -> list[Column]:
-            cols = (col for _expr in exprs for col in _expr(df))
-            return [func(cols)]
-
-        def window_function(
-            df: SparkLikeLazyFrame, window_inputs: SparkWindowInputs
-        ) -> list[Column]:
-            cols = (
-                col for _expr in exprs for col in _expr.window_function(df, window_inputs)
-            )
-            return [func(cols)]
-
-        return self._expr(
-            call=call,
-            window_function=window_function,
-            evaluate_output_names=combine_evaluate_output_names(*exprs),
-            alias_output_names=combine_alias_output_names(*exprs),
-            backend_version=self._backend_version,
-            version=self._version,
-            implementation=self._implementation,
-        )
-
     def lit(self, value: NonNestedLiteral, dtype: IntoDType | None) -> SparkLikeExpr:
         def _lit(df: SparkLikeLazyFrame) -> list[Column]:
             column = df._F.lit(value)
@@ -135,25 +110,25 @@ class SparkLikeNamespace(
         def func(cols: Iterable[Column]) -> Column:
             return reduce(operator.and_, cols)
 
-        return self._expr_from_callable(func, *exprs)
+        return self._expr.from_elementwise(func, *exprs)
 
     def any_horizontal(self, *exprs: SparkLikeExpr) -> SparkLikeExpr:
         def func(cols: Iterable[Column]) -> Column:
             return reduce(operator.or_, cols)
 
-        return self._expr_from_callable(func, *exprs)
+        return self._expr.from_elementwise(func, *exprs)
 
     def max_horizontal(self, *exprs: SparkLikeExpr) -> SparkLikeExpr:
         def func(cols: Iterable[Column]) -> Column:
             return self._F.greatest(*cols)
 
-        return self._expr_from_callable(func, *exprs)
+        return self._expr.from_elementwise(func, *exprs)
 
     def min_horizontal(self, *exprs: SparkLikeExpr) -> SparkLikeExpr:
         def func(cols: Iterable[Column]) -> Column:
             return self._F.least(*cols)
 
-        return self._expr_from_callable(func, *exprs)
+        return self._expr.from_elementwise(func, *exprs)
 
     def sum_horizontal(self, *exprs: SparkLikeExpr) -> SparkLikeExpr:
         def func(cols: Iterable[Column]) -> Column:
@@ -161,7 +136,7 @@ class SparkLikeNamespace(
                 operator.add, (self._F.coalesce(col, self._F.lit(0)) for col in cols)
             )
 
-        return self._expr_from_callable(func, *exprs)
+        return self._expr.from_elementwise(func, *exprs)
 
     def mean_horizontal(self, *exprs: SparkLikeExpr) -> SparkLikeExpr:
         def func(cols: Iterable[Column]) -> Column:
@@ -182,7 +157,7 @@ class SparkLikeNamespace(
                 ),
             )
 
-        return self._expr_from_callable(func, *exprs)
+        return self._expr.from_elementwise(func, *exprs)
 
     def concat(
         self, items: Iterable[SparkLikeLazyFrame], *, how: ConcatMethod
